ai-service/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
Pillow==10.1.0
requests==2.31.0